move yarn specifics into separate class
